separated store, added events and childAvatar views
