+++ src/Family.js
@@ -1,8 +1,8 @@
-import React, { Component } from "react";
-import { FlatList, View } from "react-native";
-import { Avatar } from "react-native-elements";
-import { fetchUsers, fetchUser, fetchRelated } from "./store";
-import { connect } from "react-redux";
+import React, { Component } from 'react';
+import { FlatList, View } from 'react-native';
+import { Avatar } from 'react-native-elements';
+import { fetchUsers, fetchUser, fetchRelated } from './store/users';
+import { connect } from 'react-redux';
 
 class Family extends Component {
   constructor() {
@@ -48,7 +48,7 @@ class Family extends Component {
       return (
         <View
           style={{
-            backgroundColor: "transparent"
+            backgroundColor: 'transparent'
           }}
         />
       );
@@ -110,18 +110,18 @@ class Family extends Component {
         <View
           style={{
             flex: 1,
-            flexDirection: "column",
-            backgroundColor: "#fff",
-            alignItems: "center",
-            justifyContent: "center"
+            flexDirection: 'column',
+            backgroundColor: '#fff',
+            alignItems: 'center',
+            justifyContent: 'center'
           }}
         >
           <View
             style={{
               flex: 1,
-              flexDirection: "row",
-              alignItems: "center",
-              justifyContent: "center"
+              flexDirection: 'row',
+              alignItems: 'center',
+              justifyContent: 'center'
             }}
           >
             <Avatar
@@ -139,7 +139,7 @@ class Family extends Component {
               }
             />
           </View>
-          <View style={{ flex: 1, flexDirection: "column" }}>
+          <View style={{ flex: 1, flexDirection: 'column' }}>
             <FlatList
               data={this.formatGrid(family, numColumns)}
               keyExtractor={this.keyExtractor}
