+++ src/SignUp.js
@@ -25,7 +25,6 @@ export default class SignUp extends Component {
   }
 
   handleSubmit = (ev, history) => {
-    // console.log(this.state);
     this.props.navigation.navigate('Family');
     this.setState({ page: 1 });
   };
