+++ src/store/events.js
@@ -1,6 +1,5 @@
 import axios from 'axios';
 
-
 const API_URL = 'https://capstone-api-server.herokuapp.com/api';
 const GET_EVENTS = 'GET_EVENTS';
 const GET_ASSIGNED = 'GET_ASSIGNED';
@@ -9,108 +8,113 @@ const CREATE_EVENT = 'CREATE_EVENT';
 const UPDATE_ASSIGNED = 'UPDATE_ASSIGNED';
 
 const getEvents = events => ({
-    type: GET_EVENTS,
-    events
+  type: GET_EVENTS,
+  events
 });
 
 const getAssigned = events => ({
-    type: GET_ASSIGNED,
-    events
+  type: GET_ASSIGNED,
+  events
 });
 
 const updateEvent = event => ({
-    type: UPDATE_EVENT,
-    event
+  type: UPDATE_EVENT,
+  event
 });
 
 const createEvent = event => ({
-    type: CREATE_EVENT,
-    event
+  type: CREATE_EVENT,
+  event
 });
 
 const updateAssigned = event => ({
-    type: UPDATE_ASSIGNED,
-    event
-})
+  type: UPDATE_ASSIGNED,
+  event
+});
 
 const eventReducer = (state = [], action) => {
-    switch (action.type) {
-        case GET_EVENTS:
-            return action.events;
-        case UPDATE_EVENT:
-            return [...state.filter(event => event.id !== action.event.id), action.event];
-        case CREATE_EVENT:
-            return [...state, action.event]
-        default:
-            return state;
-    }
-}
+  switch (action.type) {
+    case GET_EVENTS:
+      return action.events;
+    case UPDATE_EVENT:
+      return [
+        ...state.filter(event => event.id !== action.event.id),
+        action.event
+      ];
+    case CREATE_EVENT:
+      return [...state, action.event];
+    default:
+      return state;
+  }
+};
 
 const assignedEventReducer = (state = [], action) => {
-    switch (action.type) {
-        case GET_ASSIGNED:
-            return action.events;
-        case UPDATE_ASSIGNED:
-            return [...state.filter(event => event.id !== action.event.id), action.event];
-        default:
-            return state;
-    }
-}
+  switch (action.type) {
+    case GET_ASSIGNED:
+      return action.events;
+    case UPDATE_ASSIGNED:
+      return [
+        ...state.filter(event => event.id !== action.event.id),
+        action.event
+      ];
+    default:
+      return state;
+  }
+};
 
+const fetchEvents = id => {
+  return dispatch => {
+    return axios
+      .get(`${API_URL}/events/user/${id}`)
+      .then(res => {
+        return res.data;
+      })
+      .then(events => dispatch(getEvents(events)));
+  };
+};
 
-const fetchEvents = (id) => {
-    return dispatch => {
-        return axios
-            .get(`${API_URL}/events/user/${id}`)
-            .then(res => {
-                return res.data;
-            })
-            .then(events => dispatch(getEvents(events)))
-    }
-}
-
-const fetchAssigned = (id) => {
-    return dispatch => {
-        return axios
-            .get(`${API_URL}/events/assigned/${id}`)
-            .then(res => res.data)
-            .then(events => dispatch(getAssigned(events)))
-    }
-}
+const fetchAssigned = id => {
+  return dispatch => {
+    return axios
+      .get(`${API_URL}/events/assigned/${id}`)
+      .then(res => res.data)
+      .then(events => dispatch(getAssigned(events)));
+  };
+};
 
 const goUpdateAssigned = (id, updates) => {
-    return dispatch => {
-        return axios
-            .put(`${API_URL}/events/id`, updates)
-            .then(res => res.data)
-            .then(updatedEvent => dispatch(updateAssigned(updatedEvent)))
-    }
-}
+  return dispatch => {
+    return axios
+      .put(`${API_URL}/events/id`, updates)
+      .then(res => res.data)
+      .then(updatedEvent => dispatch(updateAssigned(updatedEvent)));
+  };
+};
 
 const goUpdateEvent = (id, updates) => {
-    return dispatch => {
-        return axios
-            .put(`${API_URL}/events/id`, updates)
-            .then(res => res.data)
-            .then(updatedEvent => dispatch(updateEvent(updatedEvent)))
-    }
-}
+  return dispatch => {
+    return axios
+      .put(`${API_URL}/events/id`, updates)
+      .then(res => res.data)
+      .then(updatedEvent => dispatch(updateEvent(updatedEvent)));
+  };
+};
 
-const goCreateEvent = (newEvent) => {
-    return dispatch => {
-        return axios
-            .post(`${API_URL}/events`, newEvent)
-            .then(res => res.data)
-            .then(event => dispatch(createEvent(event)))
-    }
-}
+const goCreateEvent = newEvent => {
+  return dispatch => {
+    return axios
+      .post(`${API_URL}/events`, newEvent)
+      .then(res => res.data)
+      .then(event => dispatch(createEvent(event)));
+  };
+};
 
 export {
-    fetchEvents,
-    fetchAssigned,
-    goCreateEvent,
-    goUpdateAssigned,
-    goUpdateEvent,
-    eventReducer,
-    assignedEventReducer
-}
+  fetchEvents,
+  fetchAssigned,
+  goCreateEvent,
+  goUpdateAssigned,
+  goUpdateEvent,
+  eventReducer,
+  assignedEventReducer
+};
